requirements.txt
requests==2.31.0
Pillow==9.5.0
google-api-python-client==2.108.0
duckduckgo-search==3.9.9
tqdm==4.66.1
face-recognition==1.3.0
numpy==1.24.3
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1 